(chore) Fix type errors and refactor dashboard links
This commit updates the OpenMRS dependencies and refactors the dashboard links to use the `DashboardExtension` component from the styleguide. This resolves an issue where the auto-update workflow [was failing](https://github.com/openmrs/openmrs-esm-patient-management/actions/workflows/update-openmrs-deps.yml) due to type errors related to upstream changes in how the framework handles icon properties. Changes made here include:

- Update OpenMRS dependencies to the latest `next` tagged versions
- Fix TypeScript error from framework upgrade (8.0.1-pre.3318 → 3349) where the `icon` property became more strictly typed
- Refactor appointments, service queues, ward, and patient list apps to use the `DashboardExtension` from `@openmrs/esm-styleguide` instead of local implementations
- Rename `createDashboardLink.component.tsx` to `createDashboardLink.tsx` (proper naming for factory functions)
- Remove duplicate local interfaces in favor of `Omit` utility with framework types
- Update dashboard metadata to include `basePath` and use framework types
