+++ packages/esm-patient-list-management-app/src/dashboard.meta.ts
@@ -1,5 +1,6 @@
 export const dashboardMeta = {
-  name: 'patient-lists',
+  path: 'patient-lists',
   slot: 'patient-lists-dashboard-slot',
   title: 'Patient lists',
-};
+  basePath: `${window.spaBase}/home`,
+} as const;
