--- packages/esm-service-queues-app/src/createDashboardLink.component.tsx
@@ -1,39 +1,0 @@
-import React, { useMemo } from 'react';
-import classNames from 'classnames';
-import { BrowserRouter, useLocation } from 'react-router-dom';
-import { useTranslation } from 'react-i18next';
-import { ConfigurableLink } from '@openmrs/esm-framework';
-
-export interface DashboardLinkConfig {
-  name: string;
-  title: string;
-}
-
-function DashboardExtension({ dashboardLinkConfig }: { dashboardLinkConfig: DashboardLinkConfig }) {
-  const { t } = useTranslation();
-  const { name, title } = dashboardLinkConfig;
-  const location = useLocation();
-  const spaBasePath = `${window.spaBase}/home`;
-
-  const navLink = useMemo(() => {
-    const pathArray = location.pathname.split('/home');
-    const lastElement = pathArray[pathArray.length - 1];
-    return decodeURIComponent(lastElement);
-  }, [location.pathname]);
-
-  return (
-    <ConfigurableLink
-      className={classNames('cds--side-nav__link', {
-        'active-left-nav-link': navLink.match(name),
-      })}
-      to={`${spaBasePath}/${name}`}>
-      {t('serviceQueues', 'Service queues')}
-    </ConfigurableLink>
-  );
-}
-
-export const createDashboardLink = (dashboardLinkConfig: DashboardLinkConfig) => () => (
-  <BrowserRouter>
-    <DashboardExtension dashboardLinkConfig={dashboardLinkConfig} />
-  </BrowserRouter>
-);
