+++ packages/esm-service-queues-app/src/dashboard.meta.ts
@@ -1,5 +1,6 @@
 export const dashboardMeta = {
-  name: 'service-queues',
+  path: 'service-queues',
   slot: 'service-queues-dashboard-slot',
   title: 'Service queues',
-};
+  basePath: `${window.spaBase}/home`,
+} as const;
