+++ packages/esm-ward-app/src/dashboard.meta.ts
@@ -1,0 +1,6 @@
+export const dashboardMeta = {
+  path: 'ward',
+  slot: 'ward-dashboard-slot',
+  title: 'wards',
+  basePath: `${window.spaBase}/home`,
+} as const;
