+++ packages/esm-ward-app/src/index.ts
@@ -7,7 +7,8 @@ import {
 } from '@openmrs/esm-framework';
 import { configSchema } from './config-schema';
 import { moduleName } from './constant';
-import { createDashboardLink } from './createDashboardLink.component';
+import { createDashboardLink } from './createDashboardLink';
+import { dashboardMeta } from './dashboard.meta';
 
 export const importTranslation = require.context('../translations', false, /.json$/, 'lazy');
 
@@ -19,7 +20,7 @@ const swrRefreshIntervalInMs = 60000;
 
 export const root = getAsyncLifecycle(() => import('./root.component'), options);
 
-export const wardDashboardLink = getSyncLifecycle(createDashboardLink({ name: 'ward', title: 'wards' }), options);
+export const wardDashboardLink = getSyncLifecycle(createDashboardLink(dashboardMeta), options);
 
 export const wardView = getAsyncLifecycle(() => import('./ward-view/ward-view.component'), options);
